:recycle: create abstract state for autocomplete
