+++ lib/pages/home/modals/save_post/create_post.dart
@@ -4,7 +4,6 @@ import 'package:Okuna/models/post.dart';
 import 'package:Okuna/models/post_image.dart';
 import 'package:Okuna/models/post_media.dart';
 import 'package:Okuna/models/post_video.dart';
-import 'package:Okuna/models/user.dart';
 import 'package:Okuna/pages/home/modals/save_post/widgets/create_post_text.dart';
 import 'package:Okuna/pages/home/modals/save_post/widgets/post_community_previewer.dart';
 import 'package:Okuna/pages/home/modals/save_post/widgets/post_image_previewer.dart';
@@ -27,8 +26,7 @@ import 'package:Okuna/widgets/avatars/logged_in_user_avatar.dart';
 import 'package:Okuna/widgets/avatars/avatar.dart';
 import 'package:Okuna/widgets/buttons/button.dart';
 import 'package:Okuna/widgets/buttons/pill_button.dart';
-import 'package:Okuna/widgets/contextual_search_boxes/contextual_account_search_box.dart';
-import 'package:Okuna/widgets/contextual_search_boxes/contextual_community_search_box.dart';
+import 'package:Okuna/widgets/contextual_search_boxes/contextual_search_box_state.dart';
 import 'package:Okuna/widgets/icon.dart';
 import 'package:Okuna/widgets/link_preview.dart';
 import 'package:Okuna/widgets/nav_bars/themed_nav_bar.dart';
@@ -58,7 +56,7 @@ class OBSavePostModal extends StatefulWidget {
   }
 }
 
-class OBSavePostModalState extends State<OBSavePostModal> {
+class OBSavePostModalState extends ContextualSearchBoxState<OBSavePostModal>{
   ValidationService _validationService;
   UserService _userService;
   NavigationService _navigationService;
@@ -96,14 +94,6 @@ class OBSavePostModalState extends State<OBSavePostModal> {
   bool _needsBootstrap;
   bool _isCreateCommunityPostInProgress;
 
-  TextAutocompletionService _textAccountAutocompletionService;
-  OBContextualAccountSearchBoxController _contextualAccountSearchBoxController;
-  OBContextualCommunitySearchBoxController
-      _contextualCommunitySearchBoxController;
-
-  bool _isAutocompleting;
-  TextAutocompletionType _autocompletionType;
-
   bool _isEditingPost;
 
   bool _saveInProgress;
@@ -119,17 +109,15 @@ class OBSavePostModalState extends State<OBSavePostModal> {
     _linkPreviewUrl = '';
     _isPostTextAllowedLength = false;
     _isCreateCommunityPostInProgress = false;
-    _contextualAccountSearchBoxController =
-        OBContextualAccountSearchBoxController();
-    _contextualCommunitySearchBoxController =
-        OBContextualCommunitySearchBoxController();
-    _isAutocompleting = false;
     _needsBootstrap = true;
 
     _focusNode.addListener(_onFocusNodeChanged);
   }
 
-  void _bootstrap() {
+  @override
+  void bootstrap() {
+    super.bootstrap();
+
     if (_isEditingPost) {
       _saveInProgress = false;
       _textController = TextEditingController(text: widget.post?.text ?? '');
@@ -154,6 +142,7 @@ class OBSavePostModalState extends State<OBSavePostModal> {
           text: widget.text,
           communityId: widget.community != null ? widget.community.id : null,
           draftService: _draftService);
+      setAutocompleteTextController(_textController);
       _postItemsWidgets = [
         OBCreatePostText(controller: _textController, focusNode: _focusNode)
       ];
@@ -200,12 +189,10 @@ class OBSavePostModalState extends State<OBSavePostModal> {
       _linkPreviewService = openbookProvider.linkPreviewService;
       _localizationService = openbookProvider.localizationService;
       _toastService = openbookProvider.toastService;
-      _textAccountAutocompletionService =
-          openbookProvider.textAccountAutocompletionService;
       _userService = openbookProvider.userService;
       _draftService = openbookProvider.draftService;
       _shareService = openbookProvider.shareService;
-      _bootstrap();
+      bootstrap();
       _needsBootstrap = false;
     }
 
@@ -216,20 +203,18 @@ class OBSavePostModalState extends State<OBSavePostModal> {
             child: Column(
           children: <Widget>[
             Expanded(
-                flex: _isAutocompleting ? 3 : 1,
+                flex: isAutocompleting ? 3 : 1,
                 child: Padding(
                   padding: EdgeInsets.only(left: 20.0, top: 20.0),
                   child: _buildNewPostContent(),
                 )),
-            _isAutocompleting
+            isAutocompleting
                 ? Expanded(
                     flex: 7,
-                    child: _autocompletionType == TextAutocompletionType.account
-                        ? _buildAccountSearchBox()
-                        : _buildCommunitySearchBox(),
+                    child: buildSearchBox(),
                   )
                 : const SizedBox(),
-            _isAutocompleting || (_hasImage || _hasVideo)
+            isAutocompleting || (_hasImage || _hasVideo)
                 ? const SizedBox()
                 : Container(
                     height: _hasFocus == true ? 51 : 67,
@@ -345,27 +330,6 @@ class OBSavePostModalState extends State<OBSavePostModal> {
     );
   }
 
-  Widget _buildAccountSearchBox() {
-    return OBContextualAccountSearchBox(
-      controller: _contextualAccountSearchBoxController,
-      onPostParticipantPressed: _onAccountSearchBoxUserPressed,
-      initialSearchQuery:
-          _contextualAccountSearchBoxController.getLastSearchQuery(),
-    );
-  }
-
-  Widget _buildCommunitySearchBox() {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20),
-      child: OBContextualCommunitySearchBox(
-        controller: _contextualCommunitySearchBoxController,
-        onCommunityPressed: _onCommunitySearchBoxUserPressed,
-        initialSearchQuery:
-            _contextualCommunitySearchBoxController.getLastSearchQuery(),
-      ),
-    );
-  }
-
   Widget _buildPostActions() {
     List<Widget> postActions = [];
 
@@ -448,7 +412,6 @@ class OBSavePostModalState extends State<OBSavePostModal> {
 
   void _onPostTextChanged() {
     String text = _textController.text;
-    _checkForAutocomplete();
     _checkForLinkPreview();
     setState(() {
       _charactersCount = text.length;
@@ -604,62 +567,6 @@ class OBSavePostModalState extends State<OBSavePostModal> {
     }
   }
 
-  void _checkForAutocomplete() {
-    TextAutocompletionResult result = _textAccountAutocompletionService
-        .checkTextForAutocompletion(_textController);
-
-    if (result.isAutocompleting) {
-      debugLog('Wants to autocomplete with type ${result.type} searchQuery:' +
-          result.autocompleteQuery);
-      _setIsAutocompleting(true);
-      _setAutocompletionType(result.type);
-      result.type == TextAutocompletionType.account
-          ? _contextualAccountSearchBoxController
-              .search(result.autocompleteQuery)
-          : _contextualCommunitySearchBoxController
-              .search(result.autocompleteQuery);
-    } else if (_isAutocompleting) {
-      debugLog('Finished autocompleting');
-      _setIsAutocompleting(false);
-    }
-  }
-
-  void _onAccountSearchBoxUserPressed(User user) {
-    _autocompleteFoundAccountUsername(user.username);
-  }
-
-  void _onCommunitySearchBoxUserPressed(Community community) {
-    _autocompleteFoundCommunityName(community.name);
-  }
-
-  void _autocompleteFoundAccountUsername(String foundAccountUsername) {
-    if (!_isAutocompleting) {
-      debugLog(
-          'Tried to autocomplete found account username but was not searching account');
-      return;
-    }
-
-    debugLog('Autocompleting with username:$foundAccountUsername');
-    setState(() {
-      _textAccountAutocompletionService.autocompleteTextWithUsername(
-          _textController, foundAccountUsername);
-    });
-  }
-
-  void _autocompleteFoundCommunityName(String foundCommunityName) {
-    if (!_isAutocompleting) {
-      debugLog(
-          'Tried to autocomplete found community name but was not searching community');
-      return;
-    }
-
-    debugLog('Autocompleting with community name:$foundCommunityName');
-    setState(() {
-      _textAccountAutocompletionService.autocompleteTextWithCommunityName(
-          _textController, foundCommunityName);
-    });
-  }
-
   void _setLinkPreviewUrl(String url) {
     if (_linkPreviewWidgetRemover != null) _linkPreviewWidgetRemover();
 
@@ -675,18 +582,6 @@ class OBSavePostModalState extends State<OBSavePostModal> {
     setState(() {
       _linkPreviewUrl = null;
       if (_linkPreviewWidgetRemover != null) _linkPreviewWidgetRemover();
-    });
-  }
-
-  void _setIsAutocompleting(bool isSearchingAccount) {
-    setState(() {
-      _isAutocompleting = isSearchingAccount;
-    });
-  }
-
-  void _setAutocompletionType(TextAutocompletionType autocompletionType) {
-    setState(() {
-      _autocompletionType = autocompletionType;
     });
   }
